아이템 효과가 2번 적용되는 오류 수정

+++ Assets/01.Scripts/Manager/UIManager.cs
@@ -350,7 +350,7 @@ public class UIManager : MonoSingleton<UIManager>
     }
     public void MaxHpUpdate()
     {
-        maxHpCount = Mathf.CeilToInt(GameManager.Instance.Player.playerBase.MaxHp / 4);
+        maxHpCount = Mathf.CeilToInt((float)GameManager.Instance.Player.playerBase.MaxHp / 4);
         for (int i = 0; i < maxHpCount; i++)
         {
             avcList[i].gameObject.SetActive(true);
